+++ CHANGELOG.md
@@ -6,6 +6,7 @@ but not conststraints, indexes, triggers etc.
 Use `ddlx_alter()` to create those after the data is loaded.
 - more use of explicit casts
 - `ddlx_create_trigger()` improvements by [PegoraroF10](https://github.com/PegoraroF10), now with replica and always triggers
+- change `sysid` in some places to `oid`
 
 Version 0.22
 ------------